Split tracking functionality into separate module
Restored proper code organization by separating tracking logic
from the main function handler.

Changes:
- Created tracking.py with all conversion tracking functions
- Updated __main__.py to import from tracking module
- Removed inline tracking code from __main__.py
- Cleaner separation of concerns

Structure:
- __main__.py: Core redirect logic (143 lines)
- tracking.py: Conversion tracking logic (129 lines)
- requirements.txt: Dependencies

This modular structure is now possible with the correct directory
layout and works properly with --remote-build deployment.
